Move PlayerTypes into proper interface
