+++ game-app/game-core/src/main/java/games/strategy/engine/framework/startup/launcher/ServerLauncher.java
@@ -29,7 +29,6 @@ import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
-import org.triplea.injection.Injections;
 import org.triplea.java.Interruptibles;
 import org.triplea.java.ThreadRunner;
 import org.triplea.lobby.common.GameDescription;
@@ -124,7 +123,7 @@ public class ServerLauncher implements ILauncher {
               .getGameLoader()
               .newPlayers(
                   playerListing.getLocalPlayerTypeMap(
-                      new PlayerTypes(Injections.getInstance().getPlayerTypes())));
+                      new PlayerTypes(launchAction.getPlayerTypes())));
 
       // TODO: Project#20 - if feature flag is toggled, start game relay server
       //   and use a real ClientNetworkingBridge
