+++ game-app/game-core/src/main/java/org/triplea/injection/Injections.java
@@ -1,8 +1,6 @@
 package org.triplea.injection;
 
 import com.google.common.base.Preconditions;
-import games.strategy.engine.framework.startup.ui.PlayerTypes;
-import java.util.Collection;
 import lombok.Builder;
 import lombok.Getter;
 import org.triplea.util.Version;
@@ -90,8 +88,6 @@ public final class Injections {
   @Getter private static Injections instance;
 
   private final Version engineVersion;
-  // TODO Move into LaunchAction implementation
-  private final Collection<PlayerTypes.Type> playerTypes;
 
   public static synchronized void init(final Injections injections) {
     Preconditions.checkState(getInstance() == null);
