+++ game-app/game-headed/src/main/java/org/triplea/game/client/HeadedGameRunner.java
@@ -15,14 +15,12 @@ import games.strategy.engine.framework.lookandfeel.LookAndFeel;
 import games.strategy.engine.framework.map.download.DownloadMapsWindow;
 import games.strategy.engine.framework.map.file.system.loader.ZippedMapsExtractor;
 import games.strategy.engine.framework.startup.mc.ServerModel;
-import games.strategy.engine.framework.startup.ui.PlayerTypes;
 import games.strategy.engine.framework.startup.ui.panels.main.HeadedServerSetupModel;
 import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorModel;
 import games.strategy.engine.framework.system.HttpProxy;
 import games.strategy.engine.framework.system.SystemProperties;
 import games.strategy.engine.framework.ui.MainFrame;
 import games.strategy.engine.framework.ui.background.BackgroundTaskRunner;
-import games.strategy.triplea.ai.AiProvider;
 import games.strategy.triplea.settings.ClientSetting;
 import games.strategy.triplea.ui.MacOsIntegration;
 import games.strategy.ui.Util;
@@ -34,16 +32,8 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-import java.util.ServiceLoader;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 import javax.swing.SwingUtilities;
 import lombok.extern.slf4j.Slf4j;
-import org.triplea.ai.does.nothing.DoesNothingAiProvider;
-import org.triplea.ai.flowfield.FlowFieldAiProvider;
 import org.triplea.config.product.ProductVersionReader;
 import org.triplea.debug.ErrorMessage;
 import org.triplea.injection.Injections;
@@ -135,31 +125,7 @@ public final class HeadedGameRunner {
   }
 
   private static Injections constructInjections() {
-    return Injections.builder()
-        .engineVersion(new ProductVersionReader().getVersion())
-        .playerTypes(gatherPlayerTypes())
-        .build();
-  }
-
-  private static Collection<PlayerTypes.Type> gatherPlayerTypes() {
-    return Stream.of(
-            PlayerTypes.getBuiltInPlayerTypes(),
-            List.of(
-                new PlayerTypes.AiType(new DoesNothingAiProvider()),
-                new PlayerTypes.AiType(new FlowFieldAiProvider())),
-            StreamSupport.stream(ServiceLoader.load(AiProvider.class).spliterator(), false)
-                .map(PlayerTypes.AiType::new)
-                .collect(Collectors.toSet()))
-        .flatMap(Collection::stream)
-        .filter(HeadedGameRunner::filterBetaPlayerType)
-        .collect(Collectors.toList());
-  }
-
-  private static boolean filterBetaPlayerType(final PlayerTypes.Type playerType) {
-    if (playerType.getLabel().equals("FlowField (AI)")) {
-      return ClientSetting.showBetaFeatures.getValue().orElse(false);
-    }
-    return true;
+    return Injections.builder().engineVersion(new ProductVersionReader().getVersion()).build();
   }
 
   /**
